Fix compiler warnings and compiler errors on some systems (MACOSX)
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@5907 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ module-newcamd.c
@@ -481,7 +481,7 @@ static FILTER mk_user_ftab()
   }
 
   // search CAID in client IDENT
-  cs_debug_mask(D_CLIENT, "client[%8lX].%s nfilts=%d, filt.caid=%04X", pthread_self(),
+  cs_debug_mask(D_CLIENT, "client[%8lX].%s nfilts=%d, filt.caid=%04X", (unsigned long)pthread_self(),
            cl->account->usr, cl->ftab.nfilts, filt.caid);
 
   if( !filt.caid && cl->ftab.nfilts )
